Fixes in models, set up repositories, began development of the controllers

+++ src/main/java/com/p1m5/alphastrength/model/User.java
@@ -1,19 +1,16 @@
 package com.p1m5.alphastrength.model;
 
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
+import jakarta.persistence.*;
 
 @Entity
 public class User {
 
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private Integer id;
+    private Long id;
 
     // OAuth required
-    // One-to-many relationship with workout table
+    // One-to-one relationship with workout table
     private String username;
     private String password;
 
@@ -28,7 +25,7 @@ public class User {
 
     public User() {}
 
-    public User(Integer id, String username, String password, String pictureId,
+    public User(Long id, String username, String password, String pictureId,
                 String workoutTypesStringArray, Integer height, Integer weight,
                 Integer age, String gender, String workoutPurpose) {
 
@@ -44,7 +41,7 @@ public class User {
         this.workoutPurpose = workoutPurpose;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
